Moved ReqContext into type file
